Break out mkInsertQueries
