+++ pg-toolbox/src/connection/client.ts
@@ -2,6 +2,7 @@ import migrate from 'node-pg-migrate'
 import * as pg from 'pg'
 
 import { sql, SqlQuery } from '../sql'
+import { mkInsertQueries } from './insert'
 import { MigrateOptions, parseMigrateArgs } from './migrate'
 
 export type SqlRecord = Record<string, unknown>
@@ -111,19 +112,7 @@ export class DatabaseClient {
     table: string,
     records: T[],
   ): Promise<void> {
-    const queries = records.map((record) => {
-      const columnNames = Object.keys(record)
-      const values = columnNames.map((columnName) => record[columnName])
-
-      const columnNamesSql = sql.join(columnNames.map(sql.quote), ',')
-      const valuesSql = sql.join(values.map(sql.param), ',')
-
-      return sql`
-        INSERT INTO ${sql.quote(table)} (${columnNamesSql})
-        VALUES (${valuesSql})
-      `
-    })
-
+    const queries = mkInsertQueries(table, records)
     await this.executeAll(queries)
   }
